% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who2007.r
\name{calc.zhfa}
\alias{calc.zhfa}
\title{who2007}
\usage{
calc.zhfa(mat, hfawho2007)
}
\description{
WHO 2007
Department of Nutrition for Health and Development
World Health Organization
Last modified on 08/10/2013 - Developed using R version 3.0.1 (2013-05-16)
This code concerns the  the calculation of prevalences using all vallid z-scores (non-missing)
for three indicators: weight-for-age (5 to 10 years), height-for-age (5 to 19 years) and BMI-for-age (5 to 19 years) based on the WHO 2007 references.
Exact age must be given in months (no rounding necessary), height in centimeters and weight in kilograms.
}
