% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{cohort}
\alias{cohort}
\title{Create NONMEM ready datasets for simulated clinical PK trials}
\usage{
cohort(
  data = NULL,
  include = NULL,
  n = NULL,
  obs_times = NULL,
  dose_times = NULL,
  amt = NULL,
  param = NULL,
  original_id = TRUE,
  pop_size = NULL,
  replace = FALSE,
  keep = NULL,
  tad = FALSE
)
}
\arguments{
\item{data}{A dataframe containing only one row per individual.}

\item{include}{A character string in the form of a logical R statement, to
specify the inclusion criteria for this cohort. For example,
if "WT" and "HT" are variables corresponding to weight in kg
and height in cm, respectively, you can sample only
individuals below 50kg and below 150cm by writing:
include = "WT < 50 & HT < 150"}

\item{n}{Optionally, the number of patients enroll. When using existing data,
\code{cohort()} will randomly sample \code{n} patients from your data.}

\item{obs_times}{A numeric vector of observation times.}

\item{dose_times}{A numeric vector of dosing times.}

\item{amt}{Either a numeric fixed dose, or a function that computes a
dose based on variables in your data (or which were specified
in param). If using a function, it must be Vectorized, and the
names of the variables in the function must match the names of
the variables in data (or param). This function should be
defined in the environment in which cohort is called.}

\item{param}{If creating synthetic data, this is where you specify the
distribution and parameters to use for random sampling. Supply
a list with named fields, each of which corresponds to a
variable. The value of each field should be another list,
containing (in order):
-the name of an R stats function for random sampling, e.g.
"rnorm", "runif", "rlnorm", etc.
-the arguments to the above function (except "n", which you
will have already specified).
For example, to create a normally-distributed random variable
called "WT" with mean 16.3 and standard deviation 2.5, and a
binomially distributed random variable called "HIV" with p = 0.34, write:
param = list("WT" = list("rnorm", 16.3, 2.5), "HIV" = list("rbinom", 1,
0.34))}

\item{original_id}{When \code{TRUE}, the default, \code{cohort()} will keep the same
IDs as the input data. To create new IDs starting with 1, use
\code{original_id = TRUE}.}

\item{pop_size}{Optional. When generating synthetic data, \code{cohort()} will
generate a population of size pop_size * n, and then randomly sample n
individuals from it. The default value is 10.}

\item{replace}{Optional. Whether to sample with replacement. Default: FALSE.}

\item{keep}{Optional. Character vector of column names that you do not want
converted to numeric.}

\item{tad}{Optional. Whether to calculate time after dose (TAD).}
}
\value{
A dplyr dataframe in NONMEM format.
}
\description{
The \code{cohort()} function is used to generate NONMEM-ready datasets for
clinical PK trials, either by sampling from real datasets or
generating synthetic data.
}
\section{Real Data}{

To sample from an existing dataset, pass a dataframe or file name to
\code{data}, and leave \code{param} unspecified.

Data will be filtered by the criteria specified in \code{include}, which must
use only variable names present in the data. \code{n} individuals will then be
sampled randomly.
}

\section{Synthetic Data}{

To generate synthetic data, leave \code{data} unspecified and pass the details of
the distributions from which you wish to sample to \code{param}. Currently, all
variables are assumed to be independent.
}

\section{Dosing and Observation Events}{

After sampling is finsihed, \code{cohort()} will create duplicate rows for each
individual corresponding to the timepoints specified in
\code{dose_times} and \code{obs_times}, and create a column "EVID" distinguishing
between dosing and observation events.

Finally, the function (or fixed amount) passed to \code{amt} is used to calculate
the dose for each individual at each dosing event. If using a function, it
must be Vectorized, defined in the calling environment, and its arguments
must match the names of variables used in your data. See Example 6 for
details.
}

\examples{
# 1. Sampling 20 individuals, above 10kg and below
120cm,
# with a fixed dose of 200mg, observing every 4 hours for one day and dosing
# at times 0, 5, and 12. Note that the data has columns called "WT" and "HT".

inc <- "WT > 10 & HT < 120"
ot <- seq(0, 24, by = 4)
dt <- c(0, 5, 12)

my_data <- read.csv("my_data.csv")

df1 <- cohort(my_data, include = inc, n = 20, obs_times = ot,
              dose_times = dt, amt = 200)


# 3. As in (1), except we generate new IDs.

df3 <- cohort(my_data, include = inc, n = 20, obs_times = ot,
              dose_times = dt, amt = 200, original_id = TRUE)


# 4. Simulating data. We assume WT and HT are normally distributed random
# variables, with means and standard deviations of 16 and 3.4 for WT and 132
# and 13.6 for HT.

p1 <- list("WT" = list("rnorm", 16, 3.4), "HT" = list("rnorm", 132, 13.6))

df4 <- cohort(param = p1, include = inc, n = 20,
              obs_times = ot, dose_times = dt, amt = 200)


# 5. As in (4), except we now define a dosing function.

dose_fun <- function(WT) {
  ifelse(WT < 16, 150,
  ifelse(WT < 20, 200, 250))
  }

df5 <- cohort(param = p1, include = inc, n = 20,
              obs_times = ot, dose_times = dt, amt = dose_fun)

}
\author{
Alexander Floren
}
