% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve.R
\name{curve}
\alias{curve}
\title{Detect individual PK curves}
\usage{
curve(data)
}
\arguments{
\item{data}{A data frame or data frame extension with ID, TIME, and EVID columns.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with an appended column, CURVE, containing a unique number
for each group of consecutive observations which are not separated by a non-observation
data record.
}
\description{
\code{curve()} assigns a unique number to each PK curve in a data set.
}
\details{
A PK curve is defined as any sequence of consecutive 0s in the EVID column of
length greater than or equal to 1, corresponding to the same ID.

\code{curve()} is primarily used by \code{pk_plot()} in order to draw separate lines
for distinct PK curves belonging to the same individual.
}
\author{
Sandy Floren
}
