% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etacorr.R
\name{etacorr.cont}
\alias{etacorr.cont}
\title{Plotting eta correlation for continuous and categorical variables}
\usage{
etacorr.cont(data, eta, ...)
}
\arguments{
\item{data}{Either a xpose compliant run number, nonmem data in the form of a dataframe,
or string representing the path to a nonmem tablefile}

\item{eta}{numeric vector of the desires etas}

\item{...}{unpacked, arbitrry number of symbols which represent desired covariates to plot}

\item{cat}{Default = FALSE. Set to TRUE if plotting a categorical variable. Default set to plot
continuous variables}
}
\value{
Outputs plot and returns ggplot object
}
\description{
Function that can be using to check the ditributions of eta against certain variables.
Properly coded models will have normal distributions with mean zero.
}
\examples{
etacorr(data, c(1, 2), WT, HT, AGE)
# Shows a scatterplot of WT and HT against ETA1 and ETA2 continuously.

etacorr("~/Documents/run5/simtab", c(1, 2), HIV, cat = TRUE)
# Shows a boxplot of the associated ETA1 and ETA2 with those with HIV (1) and those without (0)

}
\author{
Kendra Radtke, Dhruv Vaish, Alexander Floren
}
