% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etacorr.R
\name{etacorr_cont}
\alias{etacorr_cont}
\title{Plot ETA-covariate correlations}
\usage{
etacorr_cont(runno, eta, ...)
}
\arguments{
\item{runno}{Either a xpose compliant run number, nonmem data in the form of a dataframe,
or string representing the path to a nonmem tablefile}

\item{eta}{Which ETAs to plot, e.g. \code{eta = c(1, 2)}}

\item{...}{Covariates to plot against ETAs.}
}
\description{
For categorical covariates, use \code{etacorr_cat()}. For continuous
covariates, use \code{etacorr_cont()}.
}
\examples{
\dontrun{
# Continuous Covariates:

# Plot WT, HT, and AGE against ETA1 and ETA2 from run3
etacorr_cont(3, c(1, 2), WT, HT, AGE)

# Plot WT, HT, and AGE against ETA1 and ETA2 from an xpose_data object
xpdb <- xpose::xpose_data(3)
etacorr_cont(xpdb, c(1, 2), WT, HT, AGE)

# Categorical Covariates

# Plot HIV against ETA1 and ETA2 from run3
etacorr_cat("run3", c(1, 2), WT, HT, AGE)
}

}
\author{
Kendra Radtke, Dhruv Vaish, Sandy Floren
}
