% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_weight.R
\name{exp_weight}
\alias{exp_weight}
\title{Calculate expected weight based on age and sex for under-5 children.}
\usage{
exp_weight(age, sex, units = c("months", "years", "weeks"))
}
\arguments{
\item{age}{Age.}

\item{sex}{Sex. 1 = male, 2 = female.}

\item{units}{Units for age. Default is "months".}
}
\description{
Calculate expected weight based on age and sex for under-5 children.
}
\author{
Alexander Floren
}
