% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_triples.R
\name{find_triples}
\alias{find_triples}
\title{find_triples}
\usage{
find_triples(x, from = 4.0, to = 12.0)
}
\arguments{
\item{x:}{a vector of numeric data}

\item{from:}{a numeric lower bound for the length between two observations}

\item{to:}{a numeric upper bound for the length between two observations}
}
\value{
int representing the number of triples
}
\description{
A general-purpose function to count how many triples of numerics occur
in a vector within a fixed interval of one another.

find_triples will begin a searching for the next triple at the index of the
final member of the most recently found triple.
}
\examples{

# 1 triple
 find_triples(c(0, 4, 12, 16))

# No triples
find_triples(c(0, 12, 25))

# 2 triples
find_triples(c(0, 12, 24, 36, 48))

}
\author{
Sandy Floren
}
