% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-fmt.R
\name{round_format}
\alias{round_format}
\alias{rounde}
\title{Better number formatting}
\usage{
round_format(x, digits)

rounde(x, digits = 0)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{Number of significant significant digits to be used.}
}
\description{
Round numbers without truncating trailing zeroes.

\code{rounde()} is a rounding function that rounds up when rounding
off a 5, instead of to the nearest even number, which is what the base R
function \code{round()} does.
}
\author{
Sandy Floren

World Health Organization
}
