% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etacorr.R
\name{get_data_from_runno}
\alias{get_data_from_runno}
\title{Provide either an xpdb object, run number, or model name and return the
dataframe of results from that run.}
\usage{
get_data_from_runno(runno)
}
\description{
Provide either an xpdb object, run number, or model name and return the
dataframe of results from that run.
}
