% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_SSE.R
\name{get_sse}
\alias{get_sse}
\title{get_sse}
\usage{

}
\value{
a list of dataframes
}
\description{
Recursively search a directory for SSE results, and combine into dataframes
for parameter metrics and error.
}
\examples{

}
\author{
Alexander Floren
}
