% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_paste0.R
\name{model_paste0}
\alias{model_paste0}
\title{Flexibly process NONMEM file names}
\usage{
model_paste0(runno, ext = "")
}
\arguments{
\item{runno}{Either a NONMEM model name (with or without file extension), or
the number of a NONMEM run as an integer.}

\item{ext}{Optional. A file extension (e.g. ".lst") to append to the model
name.}
}
\description{
\code{model_paste0()} can take either an integer, the name of a model as a string,
or a complete file name and returns a model name with the desired file
extension.
}
\examples{
# Stripping a NONMEM file name to just the model name
model_paste0("run46.mod")

# Constructing a run name from a run number
model_paste0(46)

# Changing the file extension
model_paste0("run46.mod", ext = ".lst")

}
\author{
Sandy Floren
}
