% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_table.R
\name{param_table}
\alias{param_table}
\title{Create parameter tables}
\usage{

}
\arguments{
\item{...}{Run number(s) or model names(s).}

\item{nice}{Whether to use Value [\%RSE] formatting.}

\item{transform}{Whether to transform diagonal elements of OMEGA and SIGMA to standard deviations, and off-diagonal elements to correlations.}

\item{write}{Whether to write to a csv file.}

\item{filename}{Filename.}

\item{value_digits}{How many decimal places to display for parameter estimates when using "nice" formatting.}

\item{rse_digits}{How many decimal places to display for \%RSE estimates when using "nice" formatting.}

\item{max_omega}{Maximum number of OMEGAs in model.}

\item{max_sigma}{Maximum number of SIGMAs in model.}
}
\value{

}
\description{

}
\examples{

}
\author{
Sandy Floren
}
