% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pk_example.R
\docType{data}
\name{pk_example}
\alias{pk_example}
\title{PK dataset}
\format{
A tibble with 540 rows and 7 variables:
\describe{
\item{ID}{Patient ID}
\item{TIME}{Event time}
\item{AMT}{Dose amount}
\item{DV}{Dependent value (concentration)}
\item{EVID}{Event ID. 0 = observation, 1 = dose, 4 = reset-and-dose}
\item{ADDL}{number of additional doses}
\item{II}{Inter-dose interval}
}
}
\usage{
pk_example
}
\description{
This is a NONMEM-formatted dataset intended for testing savictools functions.
}
\examples{
pk_example
}
\keyword{datasets}
