% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_plot.R
\name{pk_plot}
\alias{pk_plot}
\title{Plot PK curves}
\usage{
pk_plot(data, ind = FALSE, nrow = 4, ncol = 4, id = NULL, max_tad = 26)
}
\arguments{
\item{data}{A data frame or data frame extension with ID, EVID, DV, and TIME or TAD columns.}

\item{ind}{Whether to create separate plots for each ID.}

\item{nrow}{Number of rows per page of plots when ind = TRUE.}

\item{ncol}{Number of columns per page of plots when ind = TRUE.}

\item{id}{Optional. A vector of IDs to plot, instead of plotting every ID.}

\item{max_tad}{The largest allowed value of TAD to be plotted. Defaults to 26.}
}
\value{
No return value, called for side effects
}
\description{
\code{pk_plot()} creates spaghetti PK plots with connected data points for each
PK curve.
}
\author{
Sandy Floren
}
