% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pop_example.R
\docType{data}
\name{pop_example}
\alias{pop_example}
\title{Population dataset}
\format{
A tibble with 1000 rows and 5 variables:
\describe{
\item{ID}{Patient ID}
\item{SEX}{Sex. 1 = male, 2 = female}
\item{AGE}{Age in months}
\item{WT}{Weight in kg}
\item{HT}{Event ID. 0 = observation, 1 = dose, 4 = reset-and-dose}
}
}
\usage{
pop_example
}
\description{
A dataset of demographics, with one row per individual. Used to test
savictools functions like \code{zscores()}, \code{cohort()}, and \code{expwt()}.
}
\examples{
pop_example
}
\keyword{datasets}
