% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscores.R
\name{rounde}
\alias{rounde}
\title{Implement standard rounding rules}
\usage{
rounde(x, digits = 0)
}
\arguments{
\item{x}{A numeric vector}

\item{digits}{Number of significant significant digits to round to}
}
\description{
\code{rounde()} is a rounding function that rounds up when rounding off a 5,
instead of to the nearest even number, which is what \code{round()} does.
}
