% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popPKSim.R
\name{run_sim}
\alias{run_sim}
\title{popPKSim2}
\usage{

}
\arguments{
\item{model}{TODO}

\item{dose_fun}{TODO}

\item{data}{TODO}

\item{n}{TODO}

\item{start}{TODO}

\item{end}{TODO}

\item{sample_freq}{TODO}

\item{dose_freq}{TODO}

\item{tail}{TODO}

\item{target}{TODO}

\item{output}{TODO}

\item{params}{TODO}

\item{seed}{TODO}
}
\value{

}
\description{
TODO
}
\examples{

}
\author{
Alexander Floren
}
