% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{sparse}
\alias{sparse}
\title{sparse}
\usage{
sparse(
  data,
  max_cluster = 2,
  max_distance = 24,
  col_name = "SPARSE",
  plot = TRUE,
  plot_only = FALSE,
  nrow = 4,
  ncol = 4,
  page = 1
)
}
\description{
sparse
}
