% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPC.R
\name{VPC}
\alias{VPC}
\title{Visual predictive checks}
\usage{
VPC(runno = NULL, psn_args = NULL, force = FALSE, ...)
}
\description{
\code{VPC()} allows you to run PsN's \code{vpc} command and generate a
visual predictive check with \code{xpose::xpose.VPC()} in one line of code in the
R console.
}
\examples{
  VPC(19, "-samples=500 -idv=TAD -bin_array=-0.5,0.5,1.3,2.3,3.3,5,7,10,17,28
  -bin_by_count=0", subset = "TAD < 40")



If a vpc_runXX directory already exists and force = FALSE, psn_args can be
left blank, and you only need to give the run number.





}
\author{
Alexander Floren
}
