% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscores.R
\name{zscores}
\alias{zscores}
\title{Calculate z-scores for anthrompometric indicators}
\usage{
zscores(data, missing_flag = -99, extreme_flag = -98)
}
\arguments{
\item{data}{A dataframe with ID, AGE, SEX, WT, and HT columns.}

\item{missing_flag}{Value to replace NAs. Default is -99.}

\item{extreme_flag}{Value used to replace extreme/implausible z-scores.
Default is -98.}
}
\description{
\code{zscores()} computes z-scores indicating nutritional status, adding new
columns to the end of a dataframe.

\code{zscores()} is a more streamlined implementation of the WHO scripts
\code{igrowup_standard.R}, \code{igrowup_restricted.R}, and \code{who2007.R}.
}
\author{
Kendra Radtke, Alexander Floren, Dhruv Vaish
}
