% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_plot.R
\name{cr_plot}
\alias{cr_plot}
\title{Clinical relevance plots}
\usage{
cr_plot(runno, effect_size = 0.2, lo = 0.025, hi = 0.975, ...)
}
\description{
\code{cr_plot()} visualizes the clinical relevance of covariate effects.
}
\details{
Using the variance-covariance matrix together with parameter estimates,
\code{cr_plot()} displays the posterior distributions of covariate effects
relative to the range of clinical importance.
}
