% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_plot.R
\name{cr_plot}
\alias{cr_plot}
\title{Clinical relevance plots}
\usage{
cr_plot(runno, theta, effect_size = 0.2, lo = 0.025, hi = 0.975, ...)
}
\description{
The \code{cr_plot()} function creates clinical relevance plots from NONMEM files.
Using the variance-covariance matrix together with parameter estimates,
\code{cr_plot()} shows the position of posterior distributions of covariate values
relative to the range of clinical importance (usually +/- 20\%).
}
