% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expwt.R
\name{expwt}
\alias{expwt}
\title{Calculate expected weight based on age and sex for under-5 children.}
\usage{
expwt(data, units = c("months", "years", "weeks"))
}
\arguments{
\item{data}{A dataframe with AGE and SEX columns.}

\item{units}{Units for age. Default is "months".}
}
\description{
Calculate expected weight based on age and sex for under-5 children.
}
\author{
Sandy Floren
}
