% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{nmcheck}
\alias{nmcheck}
\alias{expand_addl}
\title{NONMEM dataset utilites}
\usage{
nmcheck(data)

expand_addl(data, check = T)
}
\arguments{
\item{data}{a NONMEM-formatted dataset.}

\item{check}{whether to perform nmcheck.}
}
\description{
\code{nmcheck} checks a NONMEM dataset for formatting errors.

\code{expand_addl} expands doses coded with ADDL to distinct rows for each dose.
}
\author{
Sandy Floren
}
