% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_SSE.R
\name{parse_sse}
\alias{parse_sse}
\title{parse_sse}
\usage{

}
\arguments{
\item{summary}{}
}
\value{

}
\description{
Parse a csv file of SSE summary statistics.

To change the paramStats field, make sure you also change the corresponding
param_rows field so that they match the row numbers in your sse_results.csv.

Choose type I or type II error by changing ofv to either "I" or "II".
}
\examples{

}
\author{
Alexander Floren
}
