% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igrowup_standard.r, R/zscores.R
\name{rounde}
\alias{rounde}
\title{who2007}
\usage{
rounde(x, digits = 0)

rounde(x, digits = 0)
}
\arguments{
\item{x}{A numeric vector}

\item{digits}{Number of significant significant digits to round to}
}
\description{
WHO 2007
Department of Nutrition for Health and Development
World Health Organization
Last modified on 08/10/2013 - Developed using R version 3.0.1 (2013-05-16)
This code concerns the standard approach for the prevalences, i.e. the calculation of the prevalences takes into account all the valid (non-missing) z-scores for each of the indicators.
}
