% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.R
\name{sparse}
\alias{sparse}
\title{Detect sparse sampling occasions}
\usage{
sparse(data, min_cluster = 3, max_distance = 26, plot = FALSE)
}
\arguments{
\item{data}{A data frame or data frame extension in NONMEM format.}

\item{min_cluster}{The minimum size of a cluster of consecutive observations to be considered
intensive (i.e. not sparse) sampling.}

\item{max_distance}{The maximum allowed distance between the earliest and latest observation
of a single cluster.}

\item{plot}{Logical. Whether to make a plot of individual PK curves with sparse observations labeled.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with an appended column, SPARSE, containing a 1
for all observation points not contained in a cluster, and a 0 otherwise.
}
\description{
\code{sparse()} detects differentiates between clusters of consecutive samples
and more spread out, or "sparse," samples in a NONMEM-formatted data set.
}
\details{
For each ID, \code{sparse()} iterates through each line of the data set until
it finds an observation. It then checks for "neighbors," meaning consecutive observations without
a dose record in between. Any group of \code{min_cluster} or more observations within at most \code{max_distance} from
one another is considered a cluster.
}
\author{
Sandy Floren
}
