% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tad.R
\name{tad}
\alias{tad}
\title{Calculate time after dose (TAD)}
\usage{
tad(data)
}
\arguments{
\item{data}{A NONMEM-formatted dataframe}
}
\value{
A NONMEM-formatted dataframe with a TAD column
}
\description{
\code{tad()} computes the time after dose, adding a new column to a dataframe.
}
\details{
\code{tad()} assumes that the identifier column is called ID, the event id column
is called EVID, and the time column is called TIME. Data is grouped by ID
internally before TAD is calculated.
}
\author{
Alexander Floren
}
