% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tad.R
\name{tad}
\alias{tad}
\title{Calculate time after dose (TAD)}
\usage{
tad(data, expand = FALSE, ...)
}
\arguments{
\item{data}{A NONMEM-formatted dataframe}

\item{...}{A condition that specifies for which rows to calculate TAD.}
}
\value{
A NONMEM-formatted dataframe with a TAD column
}
\description{
\code{tad()} computes the time after dose, adding a new column to a dataframe.
}
\details{
\code{tad()} assumes that the identifier column is called ID, the event id column
is called EVID, and the time column is called TIME. Data is grouped by ID
internally before TAD is calculated.

Expressions in \code{...} are used to determine under what conditions a
row of \code{data} should count as a "dose" for calculating time after dose. This
is especially useful if there is more than one type of dose event, and TAD
should only apply to one of them. For example, suppose the DV column of
\code{data} contains concentrations of a drug \emph{X}. If a flag colum "Y_FLAG"
exists to indicate that a drug \emph{Y} is given, as opposed to \emph{X}, passing
\code{Y_FLAG = 0} to \code{...} will ensure that TAD is only calculated with respect to
drug \emph{X} dosing events.
}
\examples{
# Basic TAD calculation
tad(pk_example)

}
\author{
Sandy Floren
}
