% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VPC.R
\name{VPC}
\alias{VPC}
\title{Visual predictive checks}
\usage{
VPC(runno, psn_args = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{runno}{A run number or model name.}

\item{psn_args}{Optional. A character string containing arguments to PsN's
\code{vpc} program.}

\item{overwrite}{Optional. Logical. Should \code{VPC()} run a new VPC, deleting old
VPC directories?}

\item{...}{Optional. Additional arguments to \code{xpose4::xpose.VPC()}.}
}
\value{
No return value, called for side effects
}
\description{
\code{VPC()} allows you to run PsN's \code{vpc} command and generate a visual
predictive check with \code{xpose4::xpose.VPC()} in one line of code in the R
console.
}
\details{
Let the name of your model for VPC be called "runXX.mod". runXX.mod must
exist in your current working directory. If a vpc_runXX directory already
exists from a previous VPC, \emph{and} overwrite = FALSE, psn_args can be left
blank, and you need only supply the run number/model name.
}
\examples{
\dontrun{
# run VPC for run46.mod using runno = 46
VPC(46, "-samples=500 -bin_array=-0.5,0.5,1.5,2.5,3.5,5,7,11,13,24.5
-idv=TAD -bin_by_count=0", subset = "TAD <= 24")

# run VPC for run46.mod using runno = "run46"
VPC("run46", "-samples=500 -bin_array=-0.5,0.5,1.5,2.5,3.5,5,7,11,13,24.5
-idv=TAD -bin_by_count=0", subset = "TAD <= 24")

}
}
\author{
Sandy Floren
}
