% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscores.R
\name{zscores}
\alias{zscores}
\title{zscores}
\usage{
zscores(data, missing_flag = -99, unlikely_flag = -98)
}
\arguments{
\item{data}{A tibble of data}

\item{missing_flag}{OPTIONAL: value used to replace missing values}

\item{unlikely_flag}{OPTIONAL: value used to replace flagged z-scores}
}
\description{
Function wrapper for a script provided by the WHO to produce z-scores
}
\author{
Alexander Floren, Dhruv Vaish
}
