% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscores.R
\name{zscores}
\alias{zscores}
\title{Calculate z-scores for anthrompometric indicators}
\usage{
zscores(
  data,
  units = c("months", "years", "weeks"),
  missing_flag = NA,
  extreme_flag = NA
)
}
\arguments{
\item{data}{A dataframe with ID, AGE, SEX, WT, and HT columns.}

\item{units}{Units for AGE.}

\item{missing_flag}{Value used to replace missing z-scores Default is NA.}

\item{extreme_flag}{Value used to replace extreme/implausible z-scores.
Default is NA.}
}
\description{
\code{zscores()} computes z-scores indicating nutritional status, adding new
columns to the end of a dataframe.
}
\details{
\code{zscores()} is a more streamlined implementation of the WHO scripts
\code{igrowup_standard.R}, \code{igrowup_restricted.R}, and \code{who2007.R}.
}
